requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
charset-normalizer==3.1.0
frozenlist==1.3.3
gql==3.4.0
graphql-core==3.2.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
yarl==1.8.2
